+++ lib/AppInfo/Application.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace OCA\Reel\AppInfo;
 
 use OCA\Reel\BackgroundJob\DetectEventsJob;
+use OCA\Reel\Notification\Notifier;
 use OCP\AppFramework\App;
 use OCP\AppFramework\Bootstrap\IBootContext;
 use OCP\AppFramework\Bootstrap\IBootstrap;
@@ -23,6 +24,7 @@ class Application extends App implements IBootstrap {
     public function register(IRegistrationContext $context): void {
         // Commands are declared in appinfo/info.xml <commands>.
         // IRegistrationContext does not provide registerCommand().
+        $context->registerNotifierService(Notifier::class);
     }
 
     public function boot(IBootContext $context): void {